char-gpt/requirements.txt
torch>=2.0.0
wandb>=0.15.0
pyyaml>=6.0.0
tqdm>=4.65.0
requests>=2.31.0 